Correct transit errors
This corrects several issues with the transit routes after transfer.
Mostly, the elimination of tiny links from the network caused
some routes to look for link IDs that no longer existed. The ODX
and Alamance connector routes had to be manually corrected
given the new line network in Alamance County.