Add the all streets network
These networks reflect a lot of cleaning that took place on a separate
branch. This branch had to be deleted because the highway BIN file
was too large to commit to GitHub. To reduce the file size, I dropped
the old fields I had been carrying. These are no longer needed and
the original fields are still on the client files for reference if needed.