Simplify scenario creation checks
If the scenario is not created, you get a simple message telling you
to use the menu to create the scenario. This will apply the same to
any and all scenarios, requires no special scenario names, and
needs no maintenance moving forward (e.g. when the model is
updated to a new base year and the base scenario name changes.)