Correct directionality
Move directionality to the end of the university model. This required cleaning up the UOO macro, which converted OD back to PA (which is no longer needed).

Finally, transit trips are never converted to OD during the university model. This allows them to be added to the resident PA transit matrix.